requirements.txt
asgiref==3.10.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.6
Django==5.2.7
django-allauth==65.12.1
django-cors-headers==4.9.0
django-crispy-forms==2.4
django-environ==0.12.0
django-extensions==4.1
django-filter==25.2
djangorestframework==3.16.1
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.11
mysqlclient==2.2.7
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pillow==12.0.0
pycparser==3.0
PyJWT==2.12.1
PyMySQL==1.1.2
python-dotenv==1.2.1
python-openid==2.2.5
requests==2.33.1
sqlparse==0.5.3
urllib3==2.6.3
whitenoise==6.11.0
